Made Offline First

+++ composeApp/src/commonMain/kotlin/domain/Models/VendorPackage.kt
@@ -9,8 +9,8 @@ import kotlinx.serialization.Serializable
 @Parcelize
 data class VendorPackage(@SerialName("id") val packageId: Long = -1, @SerialName("vendor_id") val vendorId: Long = -1,
                          @SerialName("title") val title: String = "", @SerialName("description") val description: String = "",
-                         @SerialName("price") val price: Long = 0L, @SerialName("isMobileServiceAvailable") val isMobileServiceAvailable: Boolean = false,
-                         @SerialName("mobileServicePrice") val mobileServicePrice: Long = 0L, @SerialName("therapists") val packageTherapists: ArrayList<PackageTherapists> = arrayListOf(),
+                         @SerialName("price") val price: Int = 0, @SerialName("isMobileServiceAvailable") val isMobileServiceAvailable: Boolean = false,
+                         @SerialName("mobileServicePrice") val mobileServicePrice: Int = 0, @SerialName("therapists") val packageTherapists: ArrayList<PackageTherapists> = arrayListOf(),
                          @SerialName("products") val packageProducts: ArrayList<PackageProducts> = arrayListOf(), @SerialName("packageBookings") val packageBooking: Long = 0,
                          @SerialName("images") val packageImages: ArrayList<PackageImages> = arrayListOf(), @SerialName("services") val packageServices: ArrayList<PackageServices> = arrayListOf(),
                          @SerialName("reviews") val packageReviews: ArrayList<PackageReviews> = arrayListOf(), var isSelected: Boolean = false): Parcelable
